Move pointer events trap from Dropdown to Popover component
