+++ packages/components/src/dropdown/index.tsx
@@ -7,21 +7,16 @@ import type { ForwardedRef } from 'react';
 /**
  * WordPress dependencies
  */
-import { useRef, useState, useMemo, createPortal } from '@wordpress/element';
+import { useRef, useState } from '@wordpress/element';
 import { useMergeRefs } from '@wordpress/compose';
 import deprecated from '@wordpress/deprecated';
 
 /**
  * Internal dependencies
  */
-import {
-	contextConnect,
-	useContextSystem,
-	ContextSystemProvider,
-} from '../context';
+import { contextConnect, useContextSystem } from '../context';
 import { useControlledValue } from '../utils/hooks';
 import Popover from '../popover';
-import { DropdownPointerEventsCapture } from './styles';
 import type { DropdownProps, DropdownInternalContext } from './types';
 
 const UnconnectedDropdown = (
@@ -114,45 +109,6 @@ const UnconnectedDropdown = (
 		!! popoverProps?.getAnchorRect ||
 		!! popoverProps?.anchorRect;
 
-	const [ showBackdrop, setShowBackdrop ] = useState( false );
-	const contextValue = useMemo(
-		() => ( {
-			ColorPicker: {
-				onPickerDragStart() {
-					setShowBackdrop( true );
-				},
-				onPickerDragEnd() {
-					setShowBackdrop( false );
-				},
-			},
-		} ),
-		[]
-	);
-
-	const [ popoverRef, setPopoverRef ] = useState< HTMLDivElement | null >(
-		null
-	);
-	// Set the backdrop's z-index dynamically to the popover's z-index minus 1.
-	// A hardcoded fallback z-index is provided in the backdrop's styles.
-	const computedBackdropZIndexStyles = useMemo( () => {
-		if ( ! popoverRef ) {
-			return;
-		}
-
-		const popoverZIndex =
-			popoverRef.ownerDocument.defaultView?.getComputedStyle( popoverRef )
-				.zIndex;
-
-		if (
-			! popoverZIndex ||
-			Number.isNaN( Number.parseInt( popoverZIndex ) )
-		) {
-			return;
-		}
-
-		return { zIndex: Number.parseInt( popoverZIndex ) - 1 };
-	}, [ popoverRef ] );
-
 	return (
 		<div
 			className={ className }
@@ -167,52 +123,34 @@ const UnconnectedDropdown = (
 			tabIndex={ -1 }
 			style={ style }
 		>
-			{ containerRef.current && isOpen && showBackdrop
-				? /* The backdrop is rendered via portal because otherwise its fixed
-				   * position would be often affected by parent containers acting as
-				   * containing blocks.
-				   * See https://developer.mozilla.org/en-US/docs/Web/CSS/Containing_block
-				   */
-				  createPortal(
-						<DropdownPointerEventsCapture
-							aria-hidden="true"
-							onClick={ () => setShowBackdrop( false ) }
-							style={ computedBackdropZIndexStyles }
-						/>,
-						containerRef.current.ownerDocument.body
-				  )
-				: null }
-			<ContextSystemProvider value={ contextValue }>
-				{ renderToggle( args ) }
-				{ isOpen && (
-					<Popover
-						ref={ ( node ) => setPopoverRef( node ) }
-						position={ position }
-						onClose={ close }
-						onFocusOutside={ closeIfFocusOutside }
-						expandOnMobile={ expandOnMobile }
-						headerTitle={ headerTitle }
-						focusOnMount={ focusOnMount }
-						// This value is used to ensure that the dropdowns
-						// align with the editor header by default.
-						offset={ 13 }
-						anchor={
-							! popoverPropsHaveAnchor
-								? fallbackPopoverAnchor
-								: undefined
-						}
-						variant={ variant }
-						{ ...popoverProps }
-						className={ classnames(
-							'components-dropdown__content',
-							popoverProps?.className,
-							contentClassName
-						) }
-					>
-						{ renderContent( args ) }
-					</Popover>
-				) }
-			</ContextSystemProvider>
+			{ renderToggle( args ) }
+			{ isOpen && (
+				<Popover
+					position={ position }
+					onClose={ close }
+					onFocusOutside={ closeIfFocusOutside }
+					expandOnMobile={ expandOnMobile }
+					headerTitle={ headerTitle }
+					focusOnMount={ focusOnMount }
+					// This value is used to ensure that the dropdowns
+					// align with the editor header by default.
+					offset={ 13 }
+					anchor={
+						! popoverPropsHaveAnchor
+							? fallbackPopoverAnchor
+							: undefined
+					}
+					variant={ variant }
+					{ ...popoverProps }
+					className={ classnames(
+						'components-dropdown__content',
+						popoverProps?.className,
+						contentClassName
+					) }
+				>
+					{ renderContent( args ) }
+				</Popover>
+			) }
 		</div>
 	);
 };
